The Edge Scan routine is now done when the sensor is removed from its shipping jig and is mounted on MF01. There is no longer an in-jig edge scan. This reflects the current operations mode of sensor testing.
The Lateral Position routine eliminates points near the corners of the sensor where the MF07B frame is not visible. It contains corrected edge
finder steps for locating points on the protruding support frame against the background of diffuse reflection from the MF07B surface. No operator
intervention is required in finding edges in this version.